pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "screening-is-enough-pytorch"
version = "0.1.0"
description = "Paper-faithful PyTorch implementation of Screening Is Enough (Multiscreen)."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "datasets>=2.20",
  "PyYAML>=6.0",
  "schedulefree>=1.4",
  "torch>=2.5",
  "transformers>=4.52",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
